requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.32.4
beautifulsoup4==4.12.2
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.0
lxml==4.9.3
openai==1.3.0
sqlalchemy==2.0.23
mysql-connector-python==9.2.0
python-multipart==0.0.6
httpx==0.25.2